backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pypdf2==3.0.1
pdf2image==1.16.3
pillow==10.1.0
pytesseract==0.3.10
openai>=1.12.0,<2.0.0
httpx>=0.27.0,<0.28.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1

